Fix 5620754: don't show pattern screen after SIM PUK unlock
This fixes a bug where we would inadvertently show the pattern
screen after PUK-unlocking the device.  Could potentially happen
after SIM unlock as well, but that path appears to be fast enough that
it's rarely seen.

The cause was not getting the SIM state change before deciding to show
the Unlock screen.

We now immediately invoke the callback if SIM/PUK unlock succeeds without
waiting for the round-trip from the radio layer.

Change-Id: I02dcb456da415b82f30f8e3abc43f788f3931b33