stagefright aacenc: Remove a stray semicolon
This avoids warnings about an empty body in an if statement.

mem_free does the same null check, so the accidental extra
semicolon was harmless, and the whole if statement actually
is needless. But if removing the if statements, all the
other ones should be removed too, for consistency.
(This could be done as a separate change).

Change-Id: I1b6064dedd15ffa949043bb5396148aaed9b43f6