Update DroidNaskh font
Monotype has released a new version of DroidNaskh.  This version
has *no* changes in any of the glyphs.  The change is to remove
the following tables from the .ttf files.  These tables are not
used by Android and are not included in any of our other Droid
font files.  These tables are not part of the TrueType Font file
specification.  Removing these tables reduces the size of the
font files from ~370K to ~93K.  Monotype did not bump the version
because there were no user visible changes.

ASCP and TSI0 through TSI5

Change-Id: I4f9ce63b60353e2c3b6378929fba0ed9ced456e5