Fix 6119433: disambiguate screen brightness changes from other lights
This fixes a bug where the code asked to change the keyboard brightness
on a device that doesn't support it.  Instead of animating the keyboard
brightness, it ended up animating the display brightness and invoking
the power off animation as a result.  The fix is to ignore keyboard
brightness because we don't have any devices that currently support it.

Change-Id: I672d89f92f991812ea676f19c40058b2d3008656